Move BackButton component into MainTitle.jsx
