--- src/components/BackButton.tsx
@@ -1,23 +1,0 @@
-import React from "react"
-import IconButton from "@material-ui/core/IconButton"
-import ArrowBackIcon from "@material-ui/icons/ArrowBack"
-
-interface BackButtonProps {
-  onClick: () => void
-  style?: React.CSSProperties
-}
-
-const BackButton = (props: BackButtonProps) => {
-  const style = {
-    padding: 6,
-    fontSize: 32,
-    ...props.style
-  }
-  return (
-    <IconButton color="inherit" onClick={props.onClick} style={style}>
-      <ArrowBackIcon style={{ fontSize: "inherit" }} />
-    </IconButton>
-  )
-}
-
-export default BackButton
